kitti_ssc/requirements.txt
# Note that pytorch version should be 1.7.1 to be compatible with pytorch-lightning 1.4.9
# You can simply install pytorch with version 1.7.1 instead of 1.10.0 to run all 3 tasks, because this version is also compatible with the packages from mmlab.
scikit-image==0.18.1
PyYAML==5.3.1
tqdm==4.49.0
scikit-learn==0.24.0
pytorch-lightning==1.4.9
opencv-python==4.5.1.48
hydra-core==1.0.5
numpy==1.20.3
numba==0.53
imageio
protobuf~=3.19.0
torchmetrics==0.6.0

# additional packages
tqdm==4.65.0
scikit-image==0.18.1
imageio==2.28.1
tensorboard==2.13.0
scikit-learn==0.24.0
numba==0.53